+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/raw/RawNumber.kt
@@ -1,0 +1,14 @@
+package edu.nextstep.camp.calculator.domain.raw
+
+enum class RawNumber(override val char: Char) : RawExpression {
+    ZERO('0'),
+    ONE('1'),
+    TWO('2'),
+    THREE('3'),
+    FOUR('4'),
+    FIVE('5'),
+    SIX('6'),
+    SEVEN('7'),
+    EIGHT('8'),
+    NINE('9'),
+}
