+++ domain/src/main/java/edu/nextstep/camp/calculator/domain/raw/RawSign.kt
@@ -1,0 +1,8 @@
+package edu.nextstep.camp.calculator.domain.raw
+
+enum class RawSign(override val char: Char) : RawExpression {
+    PLUS('+'),
+    MINUS('-'),
+    TIMES('*'),
+    DIVISION('/'),
+}
